Add files via upload
The added files are the completed code to gather a table file in a determined place, downloading the table, calculating the desired data, and sending an email with the results.  In the file i make a statement to send an email to yourself so you can see the program working.